Parent node's information is lost when performing deepcopy.
OmegaConf misbehaves with the deepcopy function and the information
about the parent (node) is lost. This leads to errors where a variable's
value can not be resolved. The error message looks like this -
`KeyError: "str interpolation key 'hydra.ax.experiment' not found"`.
This commit provides a temporary workaround for the problem, by
setting the parent of the copied config as that of the original config.
One testcase is also added to check for this bug.