Add extra params to the method for testing class instantiation
Two new params are added to the method for testing class instantiation.
These are - key_to_get_config (for select which node in the config
should be sent to the instantiation method) and kwargs_to_pass (for the
key-value args to pass to instantiation method). Now we can merge three
of our test cases in a single function.