reimplement locate and linting
this is a reimplementation of the pydoc.locate function.  That function
would attempt to load a module from beginning to end and if any link in
the chain did not load properly then it would fail.  The reimplemented
function searches for the module to load from the end of the path to the
beginning.

Additionally, this reimplementation will raise exceptions instead of
returning None if the path cannot be located.

Signed-off-by: David Pollack <david@da3.net>