Ensure same error types are raised as before
Previously (before this PR), an ImportError was raised if a submodule
mod_a.mod_b could not be found and module and also mod_a has no
attribute named `mod_b`. An earlier commit in this PR changed this
ImportError behavior to raise an AttributeError instead. This commit
reverts that behavior -- now once again an ImportError is raised.