Copy entire `master_config` cache to `sweep_config`.
Partially reverts the intention of #749 (do not copy cached `env` resolver values), as this was made redundant by #1653 (as `env` has been deprecated, and `oc.env` does not cache values). Expanded tests to check behaviour of arbitrarily named resolvers (and not just `now` and `oc.env` resolvers) with/without caching.