rm doc on instantiate _convert_ performance
Remove the documentation paragraph about performance of `_convert_`
strategies for `instantiate`.

The paragraph made some incorrect claims, saying that `_convert_="none"`
would be the fastest strategy because no conversion takes place. The
reality is that the implementation of the `_convert_="none"` strategy
performs conversion from DictConfig to dict and then back to DictConfig,
so there will likely not be significant performance gains from using
`_convert_="none"`.

Benchmarking should be used to determine the best strategy for
performance-critical applications.