mypy: lint plugins with installed py version
In noxfile.py, run mypy on plugins using syntax from the installed
version of python (rather than the default setting, python3.7).
Additionally, pass `--follow-imports=silent` when running mypy on
plugins. This diff prevents mypy from erroring out when following
imports to type check the `numpy` package, which is installed as a
dependency of some of the plugins. Specifically, when `numpy` had been
installed with a python version >=3.8, mypy was reporting that `numpy`
had a syntax error.