Don't call next() inside assert statment in parser
`assert` statements can get removed by the Python interpreter, for instance if the env var `PYTHONOPTIMIZE` is set to 1.
Consequently, the structure of `assert next(children) ...` leads to parsing errors if the assert statment is removed, because the `children` iterator ends up in a bad state.

This PR fixes the two instances of this bug I found in `overrides_visitor.py`.
Fixes #2570.